Moved RecipeHandlerQuery to its own package-private file
